+++ src/sys/mod.rs
@@ -13,6 +13,7 @@
 mod sound; // IMPROVE
 
 pub mod arch; // Arch, *asm, detect_*, m128* m256*
+pub mod display;
 pub mod env; // App*, Arg*, Env
 pub mod fs; // Fs
 pub mod io; // Io*
@@ -31,8 +32,15 @@ crate::structural_mods! { // _mods, _pub_mods, _hidden
     }
     _pub_mods {
         pub use super::{
-            arch::_all::*, env::_all::*, fs::_all::*, io::_all::*,
-            log::_all::*, mem::_all::*, net::_all::*, os::_all::*,
+            arch::_all::*,
+            display::_all::*,
+            env::_all::*,
+            fs::_all::*,
+            io::_all::*,
+            log::_all::*,
+            mem::_all::*,
+            net::_all::*,
+            os::_all::*,
         };
         // WIPZONE
         // #[cfg(feature = "std")]