+++ docs/changelog.md
@@ -62,7 +62,7 @@ Many feature gates are removed in order to make most features make always availa
 
 
 ## features & flags
-- new features: `__publish`, `__std`, `base_safe`, `grapheme`, `safe_build`.
+- new features: `__publish`, `__std`, `base_safe`, `grapheme`, `safe_build`, `x11`.
 - remove features: `_bit*`, `_char*`, `_cmp*`, `_float_*`, `_int_*`, `_num?_all`, `_sort*`, `_str_*`, `_str_nonul`, `_str_u*`, `_text_all`, `ascii`, `cast`, `error`, `fmt`, `join`, `prim`, `split`, `str`.
 - remove flags: `bit··`, `char··`, `cmp··`, `_float··`, `_int*··`, `_nums··`, `prim··`, `sort··`, `str··`, `str_u··`.
 - add an adittional `nightly_stable_1_??` flag for the 3rd next version.
@@ -298,6 +298,10 @@ Many feature gates are removed in order to make most features make always availa
 ### arch
 - new `Arch` methods: `cntvct`, `cycles`, `rdcycle`, `rdtsc`, `rdtscp`.
 - new internal macro `ARCH!`.
+
+## display
+- new module `sys::display::x11`.
+- new types: `XDisplay`, `XError`, `XEvent`, `XWindow`.
 
 ### env
 - vendor `argv` as `IterArgSOsRef` struct and `Env` method `args_os_ref`.